% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_y_scale}
\alias{add_y_scale}
\title{Specifies the \code{y} scale of a \code{ggplot} object.}
\usage{
add_y_scale(in_g_obj, log_scale = TRUE, y_label, y_max, y_min)
}
\arguments{
\item{in_g_obj}{\code{ggplot} object to specify scale for.}

\item{log_scale}{logical vector of length 1. If true, use log scale on
\code{y}-axis; use linear scale otherwise.}

\item{y_label}{character vector of length 1. Y-axis label.}

\item{y_max}{numeric vector of length 1. Upper limit of y-axis.}
}
\value{
a \code{ggplot} object: \code{in_g_obj} with specified \code{y}-scale.
}
\description{
\code{add_y_scale} specifies the \code{y} scale of a plot.
}
\details{
Called by \code{\link{plot.results.pyro}} to unify plots.
}
