% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{arrange_grid}
\alias{arrange_grid}
\title{arrange a list of ggplot objects into a grid in a single ggplot object}
\usage{
arrange_grid(g_list, ncol)
}
\arguments{
\item{g_list}{list of ggplot objects}

\item{ncol}{if ncol = 0, this function returns the original g_list.
if ncol > 0, arrange g_list into ncol columns in a single ggplot object.}
}
\value{
either a single ggplot object or a list of ggplot objects
}
\description{
arrange a list of ggplot objects into a grid in a single ggplot object
}
