% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{calc_beta_pathway}
\alias{calc_beta_pathway}
\title{calculate viral infectivity for a given set of active pathways}
\usage{
calc_beta_pathway(pars, pathway)
}
\arguments{
\item{pars}{named list of parameter values}

\item{pathway}{"both", "endosomal", "tmprss2", "both_no_IFITM", "endosomal_no_IFITM", "tmprss2_no_IFITM"}
}
\value{
numeric vector with values of beta_E and beta_T
}
\description{
calculate viral infectivity for a given set of active pathways
}
