% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood_helpers.R
\name{calc_LL_log10normal_threshold}
\alias{calc_LL_log10normal_threshold}
\title{Calculate the log likehood of data given true model outputs if the data is log10normal distributed
and there is an observation threshold}
\usage{
calc_LL_log10normal_threshold(
  data,
  true_values,
  sd_noise,
  threshold,
  sum_LL = TRUE
)
}
\arguments{
\item{data}{numeric vector of observed data (where some number below \code{threshold}
is used to denote a measurement below the threshold)}

\item{true_values}{numeric vector of the same length as \code{data} which contains
the true model outputs}

\item{sd_noise}{numeric vector of length 1:
standard deviation of the noise distribution in log10 space}

\item{threshold}{numeric vector of length 1: observation threshold}

\item{sum_LL}{if TRUE, return sum of log likelihoods at each point; if FALSE, return
log likelihoods as vector}
}
\value{
numeric vector of length 1: the log likehood of the data given true
model outputs
}
\description{
\code{calc_LL_log10normal_threshold} calculates the log likehood of data given
true model outputs if the data is log10normal distributed
and there is an observation threshold
}
