% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{calc_moment_from_pmf}
\alias{calc_moment_from_pmf}
\title{calculate the nth moment from a probability mass function}
\usage{
calc_moment_from_pmf(x, y, n, central = FALSE)
}
\arguments{
\item{x}{numeric vector. x-values of probability mass function.  Must include
all values where the probability mass function is non-zero.}

\item{y}{numeric vector of the same length as x.  Values of probability mass
function at the points in x.}

\item{n}{integer >= 0.  Calculate nth moment.}

\item{central}{logical with default value FALSE.  If TRUE, calculate central
moment, otherwise calculate raw moment.}
}
\value{
numeric vector of length 1: the nth moment
}
\description{
calculate the nth moment from a probability mass function
}
