% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{calc_R_0_camostat}
\alias{calc_R_0_camostat}
\title{calculate R_0 for all combinations of active pathways for model 2}
\usage{
calc_R_0_camostat(pars)
}
\arguments{
\item{pars}{named list of parameter values}
}
\value{
named numeric vector with values of R_0; names are the pathways
}
\description{
calculate R_0 for all combinations of active pathways for model 2
}
