% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{calc_R_0_clare}
\alias{calc_R_0_clare}
\title{calculate R_0 for model 1}
\usage{
calc_R_0_clare(pars)
}
\arguments{
\item{pars}{named list of parameter values}
}
\value{
value of R_0
}
\description{
calculate R_0 for model 1
}
