% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{calc_r_camostat}
\alias{calc_r_camostat}
\title{calculate r for all combinations of active pathways for model 2}
\usage{
calc_r_camostat(pars, warn_if_neg = FALSE)
}
\arguments{
\item{pars}{named list of parameter values}
}
\value{
named numeric vector with values of R_0; names are the pathways
}
\description{
calculate r for all combinations of active pathways for model 2
}
