% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{calc_r_clare}
\alias{calc_r_clare}
\title{calculate r for model 1}
\usage{
calc_r_clare(pars)
}
\arguments{
\item{pars}{named list of parameter values}
}
\value{
value of r
}
\description{
calculate r for model 1
}
