% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sol.R
\name{calc_sol_chain}
\alias{calc_sol_chain}
\title{calculate solution to ODEs for each pathway for model 2, and extract median, 95\\% CI and max LL}
\usage{
calc_sol_chain(filename, model, ODE_model_filename, solving_time_max = 3)
}
\arguments{
\item{filename}{.RData file produced by lazymcmc}

\item{model}{string: "both", "endosomal", "tmprss2", "both_no_IFITM", "endosomal_no_IFITM", "tmprss2_no_IFITM"}

\item{ODE_model_filename}{name of ODE model file}

\item{solving_time_max}{time to solve until.   Default = 3 dpi.}
}
\value{
tibble with columns model, t, V, T1... and probs, which can take the values 0.025, 0.5, 0.975 or "max_LL"
}
\description{
calculate solution to ODEs for each pathway for model 2, and extract median, 95\\% CI and max LL
}
