% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sol.R
\name{calc_trajectories}
\alias{calc_trajectories}
\title{wrapper for calculating median and 95\% CI trajectories from model fits}
\usage{
calc_trajectories(filename, ODE_model_filename, two_strain = FALSE)
}
\arguments{
\item{filename}{.RData file produced by lazymcmc}

\item{ODE_model_filename}{name of ODE model file}

\item{two_strain}{TRUE for model 1, FALSE for model 2}
}
\value{
tibble of median and 95\% CI trajectories from model fits (for each pathway for model 2)
}
\description{
wrapper for calculating median and 95\% CI trajectories from model fits
}
