% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{compile_model}
\alias{compile_model}
\title{compile an odin model}
\usage{
compile_model(mod_path)
}
\arguments{
\item{mod_path}{character vector of length 1: path to ODEs}
}
\value{
compiled model
}
\description{
compile an odin model
}
