% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{CREATE_POSTERIOR_FUNC_fn_two_strain}
\alias{CREATE_POSTERIOR_FUNC_fn_two_strain}
\title{creates a closure to calculate the likelihood and model predictions for given
parameter values for model 1}
\usage{
CREATE_POSTERIOR_FUNC_fn_two_strain(parTab, data_df, PRIOR_FUNC)
}
\arguments{
\item{parTab}{data frame of parameter values created by specify_two_strain_fn}

\item{data_df}{data frame of data created by read_two_strain_data}

\item{PRIOR_FUNC}{prior function f <- function(x) 0}
}
\value{
closure to calculate likelihood and model predictions for given
parameter values
}
\description{
creates a closure to calculate the likelihood and model predictions for given
parameter values for model 1
}
