% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{darken}
\alias{darken}
\title{darken a vector of hex codes for colours}
\usage{
darken(colour_in)
}
\arguments{
\item{colour_in}{vector of hex codes for colours}
}
\value{
vector of hex codes for darkened colours
}
\description{
darken a vector of hex codes for colours
}
