% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood_helpers.R
\name{dl10norm}
\alias{dl10norm}
\title{Analogue for dlnorm but in base 10}
\usage{
dl10norm(x, mean = 1, sd = 1, log.p = FALSE)
}
\arguments{
\item{x}{a numeric vector of quantiles.}

\item{mean}{a vector of means.}

\item{sd}{a vector of standard deviations.}

\item{log.p}{a logical vector of length 1.
TRUE: return log of density,
FALSE: return density}
}
\value{
a numeric vector with the density of x in a log10normal distribution.
The log10normal distribution in log10 space has mu =  mean and sigma = sd.
}
\description{
\code{dl10norm} is an analogue for dlnorm but in base 10
}
