% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{extract_max_LL_iters}
\alias{extract_max_LL_iters}
\title{given an RData file, load the associated MCMC chains and find the iterations with maximum likelihood}
\usage{
extract_max_LL_iters(filename_in, n_par_sets)
}
\arguments{
\item{filename_in}{character vector of length 1: name of .RData file outputted by MCMC}

\item{n_par_sets}{numeric vector of length 1: find the n_par_sets parameter sets
with the highest likelihood}
}
\value{
data frame with iterations as rows
}
\description{
given an RData file, load the associated MCMC chains and find the iterations with maximum likelihood
}
