% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{find_sign_change_ind}
\alias{find_sign_change_ind}
\title{find the indices at which a vector changes sign}
\usage{
find_sign_change_ind(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a logical vector of length length(x) - 1.  TRUE at element i indicates
that x\link{i} has a different sign from x\link{i+1}.
}
\description{
find the indices at which a vector changes sign
}
