% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{gen_default_mcmcPars}
\alias{gen_default_mcmcPars}
\title{generate default MCMC parameters}
\usage{
gen_default_mcmcPars(
  iterations = 10000,
  popt = 0.44,
  opt_freq = 1000,
  thin = 1,
  adaptive_period = 5000,
  save_block = 100
)
}
\arguments{
\item{iterations}{numeric vector of length 1: number of iterations for which to run MCMC (excluding adaptive period)}

\item{popt:}{numeric vector of length 1: optimum acceptance ratio for adapting step size}

\item{opt_freq:}{numeric vector of length 1: adjust step size per this many number of iterations during adaptive period}

\item{thin:}{numeric vector of length 1: save state of chain and parameter values to file every this many iterations}

\item{adaptive_period:}{numeric vector of length 1: number of iterations for which to adapt MCMC}

\item{save_block:}{numeric vector of length 1: number of iterations to hold in memory}
}
\value{
a named vector with the above elements
}
\description{
\code{gen_default_mcmcPars} supplies some default MCMC parameters.
}
\details{
All arguments are optional (to overwrite defaults)
}
