% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{gen_mcmcPars_short}
\alias{gen_mcmcPars_short}
\title{generate default MCMC parameter values for a short chain}
\usage{
gen_mcmcPars_short(temperature = 1, parallel_tempering_iter)
}
\arguments{
\item{temperature}{numeric vector of temperatures for parallel tempering}

\item{parallel_tempering_iter}{numeric vector of length 1: number of
iterations before attempting to swap adjacent chains}
}
\value{
list of default parameter values for a short chain
}
\description{
generate default MCMC parameter values for a short chain
}
