% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen_nice_breaks}
\alias{gen_nice_breaks}
\title{format a numeric vector of breaks nicely for printing}
\usage{
gen_nice_breaks(breaks, limits)
}
\arguments{
\item{breaks}{numeric vector}

\item{limits}{numeric vector of length 2. If present, breaks must be within limits}
}
\value{
list with two elements:
breaks: numeric vector of same length
break_names: character vector of same length
order_magnitude: highest order of magnitude of breaks
}
\description{
format a numeric vector of breaks nicely for printing
}
