% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{gen_prediction_ci}
\alias{gen_prediction_ci}
\title{produce confidence intervals for model predictions}
\usage{
gen_prediction_ci(chain, prediction_names, prediction_times, samples = FALSE)
}
\arguments{
\item{samples}{logical.  if TRUE, generate 20 samples instead of 95\% CIs.}

\item{data_df}{data frame inputted into MCMC inference}

\item{chain:}{data frame resulting from reading in csv file outputted from \code{run_mcmc}
should discard burn-in before running this function.}

\item{prediction_names:}{character vector: column names in \code{chain} which
contain the model predictions}

\item{sampling_time:}{numeric vector: times at which data is sampled}
}
\value{
a data frame which consists of columns containing the CIs
for the model predictions at the sampling times (padded with NAs if necessary)
}
\description{
\code{gen_prediction_ci} produces confidence intervals for model predictions
}
