% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{gen_summary_statistics_fn_two_strain}
\alias{gen_summary_statistics_fn_two_strain}
\title{create closure to calculate summary statistics for model 1 and provide bounds for them}
\usage{
gen_summary_statistics_fn_two_strain(parTab, strains)
}
\arguments{
\item{parTab}{data frame}
}
\value{
a list with elements
"calc_summary": closure to calculate
}
\description{
create closure to calculate summary statistics for model 1 and provide bounds for them
}
