% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{gen_time_list}
\alias{gen_time_list}
\title{Generate vector of times at which to solve ODE for fitting}
\usage{
gen_time_list(sampling_time)
}
\arguments{
\item{sampling_time}{a numeric vector of length n: times at which we have data}
}
\value{
a list with the elements
mcmc_solving_time: a numeric vector: if sampling time includes 0, mcmc_solving_time = sampling_time,
otherwise append 0
sampling_ind: a numeric vector containing the indices of mcmc_solving_time which are in sampling_time
}
\description{
\code{gen_time_list} generates a vector of times at which to solve ODE for fitting
}
