% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{get_max_LL_params}
\alias{get_max_LL_params}
\title{given an RData file, load the associated MCMC chains and find the parameters with maximum likelihood}
\usage{
get_max_LL_params(filename_in)
}
\arguments{
\item{filename_in}{character vector of length 1: name of .RData file outputted by MCMC}
}
\value{
vector of parameter values
}
\description{
given an RData file, load the associated MCMC chains and find the parameters with maximum likelihood
}
