% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R, R/plotting.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulates the ggplot colour palette.}
\usage{
gg_color_hue(n)

gg_color_hue(n)
}
\arguments{
\item{n}{numeric vector of length 1. The number of colours required.}
}
\value{
a character vector of length \code{n}. Each element is a hex colour code.

a character vector of length \code{n}. Each element is a hex colour code.
}
\description{
\code{gg_color_hue} emulates the ggplot colour palette.

\code{gg_color_hue} emulates the ggplot colour palette.
}
