% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{ggsave_wch}
\alias{ggsave_wch}
\title{version of ggsave which suppresses warnings that rows are removed from the data frame,
and about transformations introducing infinite values}
\usage{
ggsave_wch(...)
}
\description{
version of ggsave which suppresses warnings that rows are removed from the data frame,
and about transformations introducing infinite values
}
