% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{ggsave_wrapper}
\alias{ggsave_wrapper}
\title{wrapper function to feed arguments to ggsave}
\usage{
ggsave_wrapper(g, filename_fn, width, height, filename_args)
}
\arguments{
\item{g}{either a ggplot object, or a list of ggplot objects of length m,
or a nested list of ggplot objects (length n, each of length m)}

\item{filename_fn}{function which makes a filename by taking n arguments and
outputting a character vector of length 1. If g is a ggplot object or a list
of ggplot objects of length m, then n = 1}

\item{width}{numeric vector of length 1. width of figure in cm}

\item{height}{numeric vector of length 1. height of figure in cm}

\item{filename_args}{if g is a ggplot object, a charater vector of length 1.
if g is a list of ggplot objects of length m, a character vector of length m.
if g is a nested list of ggplot objects, a list of length 2, containing a
character vector of length n and a character vector of length m respectively.}
}
\description{
wrapper function to feed arguments to ggsave
}
