% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{calculate geometric mean of a vector}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical vector of length 1}
}
\value{
numeric vector of length 1: geometric mean
}
\description{
calculate geometric mean of a vector
}
