% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{is_integer_like}
\alias{is_integer_like}
\title{check if x is within \code{tol} of \code{x}}
\usage{
is_integer_like(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{numeric vector of length 1}

\item{tol}{tolerance}
}
\value{
logical vector of length 1
}
\description{
check if x is within \code{tol} of \code{x}
}
