% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lighten}
\alias{lighten}
\title{lighten a vector of hex codes for colours}
\usage{
lighten(colour_in)
}
\arguments{
\item{colour_in}{vector of hex codes for colours}
}
\value{
vector of hex codes for lighened colours
}
\description{
lighten a vector of hex codes for colours
}
