% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{load_into_list}
\alias{load_into_list}
\title{load contents of RData file into a list}
\usage{
load_into_list(filename, var_names)
}
\arguments{
\item{filename}{character vector of length 1: RData filename}

\item{var_names}{optional argument specifying variables to load.  If missing,
load all variables}
}
\value{
list containing variables with names var_names
}
\description{
load contents of RData file into a list
}
