% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{load_packages}
\alias{load_packages}
\title{loads a character vector of packages}
\usage{
load_packages(package_vector)

load_packages(package_vector)
}
\arguments{
\item{package_vector:}{character vector to load}
}
\description{
loads a character vector of packages

loads a character vector of packages
}
