% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_factor}
\alias{make_factor}
\title{make factor out of a numeric vector}
\usage{
make_factor(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
a character version of the vector, with levels in numeric order
}
\description{
make factor out of a numeric vector
}
