% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_filename}
\alias{make_filename}
\title{construct plot filename}
\usage{
make_filename(filename, par_name, description)
}
\arguments{
\item{filename}{character vector of length 1. experiment/chain identifier}

\item{par_name}{character vector of length 1. parameter name}

\item{description}{character vector of length 1. description of plot}
}
\value{
character vector of length 1.
}
\description{
construct plot filename
}
