% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_filename_from_number}
\alias{make_filename_from_number}
\title{make string from number in format acceptable for filename}
\usage{
make_filename_from_number(x, decimal_points = 3, scientific = FALSE)
}
\arguments{
\item{x}{numeric vector of length 1}

\item{decimal_points}{numeric vector of length 1. number of decimal points to
round to}

\item{scientific}{logical vector of length 1.  whether to use scientific
notation}
}
\value{
character vector of lenth 1: string suitable for use as filename
}
\description{
\code{make_filename_from_number} makes a string from a number, replacing
periods and plus signs
}
