% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_p_value_str}
\alias{make_p_value_str}
\title{format a p-value for plotting}
\usage{
make_p_value_str(p_value)
}
\arguments{
\item{p-value}{numeric vector of length 1}
}
\value{
character vector of length 1
}
\description{
format a p-value for plotting
}
