% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_seed}
\alias{make_seed}
\title{make a default list of random number generator seeds for MCMC}
\usage{
make_seed(n_replicates, n_temperatures, seed_start = 1)
}
\arguments{
\item{n_replicates}{the number of parallel chains to run to assess convergence}

\item{n_temperatures}{the number of temperatures}

\item{seed_start}{integer to start seed sequence}
}
\value{
an n_replicates by n_temperatures of integers
}
\description{
make a default list of random number generator seeds for MCMC
}
