% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_common_fns.R
\name{make_solving_time}
\alias{make_solving_time}
\title{make a vector of solving times from a vector of times}
\usage{
make_solving_time(times_in)
}
\arguments{
\item{times_in}{numeric vector of times}
}
\value{
numeric vector: sorted unique times, with t = 0
}
\description{
make a vector of solving times from a vector of times
}
