% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{mkdir_from_filename}
\alias{mkdir_from_filename}
\title{makes a directory from a filename}
\usage{
mkdir_from_filename(filename)
}
\arguments{
\item{filename}{character vector of length 1}
}
\description{
makes a directory from a filename
}
