% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{no_warn}
\alias{no_warn}
\title{a version of a function that suppresses warnings}
\usage{
no_warn(f)
}
\arguments{
\item{f}{a function}
}
\value{
the same function without warnings
}
\description{
a version of a function that suppresses warnings
}
