% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood_helpers.R
\name{pl10norm}
\alias{pl10norm}
\title{Analogue for plnorm but in base 10}
\usage{
pl10norm(q, mean = 1, sd = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{a numeric vector of quantiles.}

\item{mean}{a vector of means.}

\item{sd}{a vector of standard deviations.}

\item{lower.tail}{TRUE: calculate the probability of drawing a number less than q.
FALSE: = calculate the probability of drawing a number larger than q.}

\item{log}{a logical vector of length 1.
TRUE: return log of probability,
FALSE: return probability}
}
\value{
a numeric vector with the probability of drawing a number less than
(or greater than, depending on the value of \code{lower.tail})
in a log10normal distribution.
The log10normal distribution in log10 space has mu =  mean and sigma = sd.
}
\description{
\code{pl10norm} is an analogue for plnorm but in base 10
}
