% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate_scatter}
\alias{plot_bivariate_scatter}
\title{make scatter plot of samples from joint posterior distribution across two parameters}
\usage{
plot_bivariate_scatter(
  chain,
  true_value,
  start_value,
  var_name,
  lower_bounds,
  upper_bounds,
  labels,
  n_samples = 0,
  zoom = FALSE
)
}
\arguments{
\item{chain}{data table containing parameter values}

\item{true_value}{If simulated data is used, numeric vector of length 2
containing the true parameter values, otherwise c(NA, NA)}

\item{start_value}{nx2 matrix of starting parameter values, where n is the
number f parallel MCMC chains}

\item{var_name}{character vector of length 2.  Parameter names in colnames(chain)}

\item{lower_bounds}{numeric vector of length 2. Lower axis limits.}

\item{upper_bounds}{numeric vector of length 2.  Upper axis limits.}

\item{labels}{character vector of length 2. Parameter names to display on plot}

\item{n_samples}{numeric vector of length 1.  number of samples to plot, taken
at evenly spaced intervals along the chain.  if 0, plot all samples.}

\item{zoom}{logical vector of length 1. if TRUE, the axis limits are the
range of the parameter values in the chain (the start value is disregarded).
if FALSE, the limits are as given in lower_bounds and upper_bounds,
unless at least one of these is NA,
in which case the range of the chain values + start value is used.}
}
\value{
a ggplot object
}
\description{
make scatter plot of samples from joint posterior distribution across two parameters
}
