% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate_scatter_all}
\alias{plot_bivariate_scatter_all}
\title{plot bivariate scatter plots of all fitted variables}
\usage{
plot_bivariate_scatter_all(
  chain,
  parTab,
  startTab,
  ncol = 1,
  n_samples = 0,
  zoom,
  real_data = real_data
)
}
\arguments{
\item{chain}{data table containing MCMC iterations after discarding burn-in}

\item{parTab}{data frame containing parameter information constructed by
specify_parameters()}

\item{startTab}{either a data frame of the same form as parTab, containing
starting parameter values (before discarding burn-in), or a list of such data
frames, one for each parallel chain}

\item{ncol}{if ncol = 0, output each plot as a separate ggplot object.
if ncol > 0, arrange into ncol columns in a single ggplot object.}

\item{n_samples}{numeric vector of length 1.  number of samples to plot, taken
at evenly spaced intervals along the chain.  if 0, plot all samples.}

\item{zoom}{logical vector of length 1. if TRUE, the axis limits are the
range of the parameter values in the chain (the start value is disregarded).
if FALSE, the limits are as given in parTab,
unless at least one of these is NA,
in which case the range of the chain values + start value is used.}

\item{real_data}{logical vector ot length 1.  If TRUE, because we're using
real data, we don't know the true parameter values, so the values in parTab
are just example values.  If FALSE, we are using simulated data, and the
values in parTab are the true values, so plot these}
}
\value{
ggplot object or a list of ggplot objects
}
\description{
plot bivariate scatter plots of all fitted variables
}
