% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{plot the potential scale reduction factor vs the number of MCMC iterations}
\usage{
plot_diagnostics(max_psrf, iterations)
}
\arguments{
\item{max_psrf}{numeric vector.  The maximum PSRF among the parameters each
time the PSRF is calculated}

\item{iterations}{numeric voector of length 1. the potential scale reduction
factor is calculated once every this many iterations}
}
\value{
ggplot object
}
\description{
plot the potential scale reduction factor vs the number of MCMC iterations
}
