% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dynamics_inner}
\alias{plot_dynamics_inner}
\title{plot virus dynamics}
\usage{
plot_dynamics_inner(
  data_df,
  ci_df,
  samples_df,
  noise_var,
  true_var,
  y_label,
  error_string,
  y_min,
  y_max,
  obs_threshold,
  x_label = "Time (hpi)"
)
}
\value{
ggplot object
}
\description{
plot virus dynamics
}
