% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_histogram_general}
\alias{plot_histogram_general}
\title{plot a nicer-looking histogram}
\usage{
plot_histogram_general(
  vec,
  special_value,
  breaks,
  vline_values,
  limits,
  plot_title,
  plot_xlab
)
}
\arguments{
\item{vec}{numeric vector of valuse from which to make histogram}

\item{special_value}{optional parameter.  If specified, should be numeric
vector of length 1. Plot the bar for this value in a different colour. Used,
for example, to draw attention to the bar at x = 0.}

\item{breaks}{optional parameter.  If specified, should be numeric vector.
Specifies breaks on horizontal axis.  If not given, use horizontal axis
limits as breaks.}

\item{vline_values}{optional parameter.  If specified, should be numeric
vector. Plot vertical lines at these values.}

\item{limits}{optional parameter. If specified, should be numeric vector of
length 2.  Use as horizontal axis limits. If not given use the range of vec
(if there are values below 0) or \link{0, max(vec)} (if there are no values below
zero).}

\item{plot_title}{optional parameter. If specified, should be character
vector of length 1. Title of plot.}

\item{plot_xlab}{character vector of length 1. Horizontal axis label.}
}
\value{
ggplot object
}
\description{
plot a nicer-looking histogram
}
