% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_LL}
\alias{plot_LL}
\title{plot trace of log likelihood for each chain}
\usage{
plot_LL(filename)
}
\arguments{
\item{filename}{string containing filename of .RData file outputted by lazymcmc}
}
\value{
ggplot object
}
\description{
plot trace of log likelihood for each chain
}
