% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_marginal_by_chain}
\alias{plot_marginal_by_chain}
\title{plot marginal histogram, either coloured or faceted by chain}
\usage{
plot_marginal_by_chain(dir_name, par_name)
}
\arguments{
\item{dir_name}{string: directoy where results are saved}

\item{par_name}{string: parameter to plot}
}
\value{
list of 2 ggplot objects.  First is marginal histogram coloured by chain,
second is marginal histogram faceted by chain
}
\description{
plot marginal histogram, either coloured or faceted by chain
}
