% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_marginal_histogram}
\alias{plot_marginal_histogram}
\title{plot the marginal histogram of a single fitted variable}
\usage{
plot_marginal_histogram(chain, x_lab, xlim)
}
\arguments{
\item{chain}{a single row of a data table containing the parameter values}

\item{x_lab}{parameter name to display on x-axis}

\item{xlim}{numeric vector of length 2.  limits of x-axis}
}
\value{
ggplot object
}
\description{
plot the marginal histogram of a single fitted variable
}
