% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_marginal_histogram_all}
\alias{plot_marginal_histogram_all}
\title{plotting the marginal histogram of all fitted variables}
\usage{
plot_marginal_histogram_all(chain, parTab, ncol = 1)
}
\arguments{
\item{chain}{data table containing iterations of MCMC chain}

\item{parTab}{data frame containing parameters}

\item{ncol}{if ncol = 0, output each plot as a separate ggplot object.
if ncol > 0, arrange into ncol columns in a single ggplot object.}
}
\value{
ggplot object or a list of ggplot objects
}
\description{
plotting the marginal histogram of all fitted variables
}
