% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sol.R
\name{plot_trajectories_camostat_amphoB}
\alias{plot_trajectories_camostat_amphoB}
\title{plot model 2 fits}
\usage{
plot_trajectories_camostat_amphoB(strain1, Calu3 = FALSE, trajectory_filename)
}
\arguments{
\item{strain1}{"omicron" or "Delta"}

\item{Calu3}{if TRUE plot Calu3 trajectories; if FALSE plot hNEC}

\item{trajectory_filename}{filename which trajectories were saved in}
}
\value{
filenames which plots are saved to
}
\description{
plot model 2 fits
}
