% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_postprocessing.R
\name{postprocess_chain}
\alias{postprocess_chain}
\title{more postprocessing of MCMC output}
\usage{
postprocess_chain(
  chain,
  parTab,
  data_df,
  bind_summary_to_chain_partial,
  calc_residuals,
  get_prediction_ci_df,
  filename,
  combined_flag
)
}
\arguments{
\item{chain}{data frame containing MCMC chain}

\item{parTab}{data frame of model parameters created by specify_parameters}

\item{bind_summary_to_chain_partial}{closure created by postprocess_for_plotting
used to calculate summary statistics and bind them to chain}

\item{calc_residuals}{closure. NULL or function to calculate residuals}

\item{get_prediction_ci_df}{closure. NULL or function to calculate credible
intervals for model predictions}

\item{filename}{character vector of length 1: name of .RData file outputted by MCMC}

\item{combined_flag}{logical.  INdicates whether this chain is combined from
parallel chains that have converged}

\item{data}{data frame of data used for fitting in MCMC}
}
\description{
more postprocessing of MCMC output
}
