% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_postprocessing.R
\name{postprocess_for_plotting}
\alias{postprocess_for_plotting}
\title{postprocessing of MCMC output}
\usage{
postprocess_for_plotting(
  filename,
  run_parallel = TRUE,
  gen_summary_statistics = NULL,
  calc_residuals = NULL,
  get_prediction_ci_df = NULL,
  plot_flags
)
}
\arguments{
\item{filename}{character vector of length 1: name of .RData file outputted by MCMC}

\item{run_parallel}{logical. Whether to run calculations in parallel}

\item{gen_summary_statistics}{closure.  NULL or function to calculate summary statistics}

\item{calc_residuals}{closure. NULL or function to calculate residuals}

\item{get_prediction_ci_df}{closure. NULL or function to calculate credible
intervals for model predictions}
}
\description{
postprocessing of MCMC output
}
