% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{read_csv_or_rds}
\alias{read_csv_or_rds}
\title{checks if file exists in rds or csv format, and reads rds file preferentially}
\usage{
read_csv_or_rds(filename, ...)
}
\arguments{
\item{filename}{character vector of length 1: with or without extension}
}
\value{
the read data frame
}
\description{
checks if file exists in rds or csv format, and reads from original extension if given;
else reads rds file preferentially
}
