% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read data}
\usage{
read_data(strain1, Calu3)
}
\arguments{
\item{strain1}{"omicron" or "delta"}

\item{Calu3}{logical. if TRUE, read Calu3 data, if FALSE, read hNEC data}
}
\value{
a data frame
}
\description{
read data
}
