% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{read_prism_file}
\alias{read_prism_file}
\title{construct a data frame from a prism file}
\usage{
read_prism_file(filename)
}
\arguments{
\item{filename}{character vector of length 1.  filename of prism file}
}
\value{
data frame of data in prism file
}
\description{
construct a data frame from a prism file
}
