% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{read_summary_stats}
\alias{read_summary_stats}
\title{reads summary stats from a tex file generated by running MCMC}
\usage{
read_summary_stats(filename, param_name)
}
\arguments{
\item{filename}{character vector of length 1}

\item{param_name}{name of parameter for which to retrieve summary stats}
}
\value{
double vector containing summary stats
}
\description{
reads summary stats from a tex file generated by running MCMC
}
