% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_two_strain_data}
\alias{read_two_strain_data}
\title{read data for two strains}
\usage{
read_two_strain_data(strains, Calu3)
}
\arguments{
\item{strains}{vector of strain names}

\item{Calu3}{logical. if TRUE, read Calu3 data, if FALSE, read hNEC data}
}
\value{
a data frame
}
\description{
read data for two strains
}
