% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_exp_camostat_amphoB}
\alias{run_exp_camostat_amphoB}
\title{fit model 2 to data}
\usage{
run_exp_camostat_amphoB(
  save_dir,
  strain,
  Calu3,
  tau_E = 4,
  tau_T = 4,
  length_run = 2,
  run_flag = TRUE
)
}
\arguments{
\item{save_dir}{string ending in "/": directory to save results in}

\item{strain}{character vector.  name of strain.}

\item{Calu3}{logical. if TRUE, fit model to Calu3 data, if FALSE, use hNEC data}

\item{tau_E}{fixed value of tau_E}

\item{tau_T}{fixed value of tau_T}

\item{length_run}{integer.  If length == 1, short run (for testing purposes); if length == 2, normal run; if length = 3, extra long run.
otherwise, run until convergence}

\item{run_flag}{logial.  if TRUE, run and postprocess; if FALSE, postprocess only}
}
\value{
NULL (results saved to file)
}
\description{
fit model 2 to data
}
