% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_exp_two_strain}
\alias{run_exp_two_strain}
\title{fit model 1 to data}
\usage{
run_exp_two_strain(save_dir, Calu3, length_run = 2, run_flag = TRUE)
}
\arguments{
\item{save_dir}{string ending in "/": directory to save results in}

\item{Calu3}{logical. if TRUE, fit model to Calu3 data, if FALSE, use hNEC data}

\item{length_run}{integer.  If length == 1, short run (for testing purposes); if length == 2, normal run; if length = 3, extra long run.
otherwise, run until convergence}

\item{run_flag}{logical.  if TRUE, run and postprocess; if FALSE, postprocess only}
}
\value{
NULL (results saved to file)
}
\description{
fit model 1 to data
}
