% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_postprocessing.R
\name{sample_from_prior}
\alias{sample_from_prior}
\title{samples from joint prior distribution}
\usage{
sample_from_prior(
  PRIOR_FUNC,
  parTab,
  startTab,
  data_df,
  filenames,
  run_parallel
)
}
\arguments{
\item{PRIOR_FUNC}{NULL for uniform prior, otherwise closure specifying prior}

\item{startTab}{data frame of smae format as parTab, storing starting
parameter values}

\item{filenames}{vector of filenames in which to temporarily store samples}

\item{run_parallel}{logical. Whether to run calculations in parallel}
}
\value{
data frame of samples from the prior distribution
}
\description{
samples from joint prior distribution
}
