% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{save_all}
\alias{save_all}
\title{saves all variables in parent environment}
\usage{
save_all(filename)
}
\arguments{
\item{filename}{character vector of length 1: no extension}
}
\description{
same as save.image except for parent environment rather than global
}
