% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{set_default_seeds}
\alias{set_default_seeds}
\title{set default random number generator seed for each}
\usage{
set_default_seeds(n_replicates, n_temperatures)
}
\description{
set default random number generator seed for each
}
