% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{source_files}
\alias{source_files}
\title{sources a character vector of files}
\usage{
source_files(source_vector)

source_files(source_vector)
}
\arguments{
\item{source_vector:}{character vector of files to source}
}
\description{
sources a character vector of files

sources a character vector of files
}
