% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_pars.R
\name{specify_clare_IFITM_fn}
\alias{specify_clare_IFITM_fn}
\title{make parameter table for model 2}
\usage{
specify_clare_IFITM_fn(Calu3, tau_E, tau_T)
}
\arguments{
\item{Calu3}{if TRUE, parameter values for Calu-3 cells; if FALSE, values for hNECs}

\item{tau_E}{value of tau_E}

\item{tau_T}{value of tau_T}
}
\value{
data_df frame with columns
values: default parameter value
names: internal name for parameter
fixed:  if 0, fit parameter, otherwise fix parameter
lower_bound: lower bound of parameter for fitting
upper_bound: upper bound of parameter for fitting
steps: width of proposal distribution in parameter space
names_plot: character vector which which to label parameter distribution plot
}
\description{
make parameter table for model 2
}
