% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_pars.R
\name{specify_clare_parameters_fn}
\alias{specify_clare_parameters_fn}
\title{make parameter table}
\usage{
specify_clare_parameters_fn(Calu3)
}
\arguments{
\item{Calu3}{if TRUE, parameter values for Calu-3 cells}
}
\value{
data_df frame with columns
values: default parameter value
names: internal name for parameter
fixed:  if 0, fit parameter, otherwise fix parameter
lower_bound: lower bound of parameter for fitting
upper_bound: upper bound of parameter for fitting
steps: width of proposal distribution in parameter space
names_plot: character vector which which to label parameter distribution plot
}
\description{
make parameter table
}
