% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_pars.R
\name{specify_two_strain_fn}
\alias{specify_two_strain_fn}
\title{make parameter table for model 1}
\usage{
specify_two_strain_fn(strains, Calu3)
}
\arguments{
\item{strains}{vector of strain names}

\item{Calu3}{if TRUE, parameter values for Calu-3 cells; if FALSE, values for hNECs}
}
\value{
data_df frame with columns
values: default parameter value
names: internal name for parameter
fixed:  if 0, fit parameter, otherwise fix parameter
lower_bound: lower bound of parameter for fitting
upper_bound: upper bound of parameter for fitting
steps: width of proposal distribution in parameter space
names_plot: character vector which which to label parameter distribution plot
}
\description{
make parameter table for model 1
}
