% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{sum_stages}
\alias{sum_stages}
\title{sum time series of odin solution across compartments defined as vectors}
\usage{
sum_stages(df, compartment_names)
}
\arguments{
\item{df}{data frame containing odin solution}

\item{compartment_names}{character vector. Names of compartments to sum over}
}
\value{
a data frame with one column for each element of compartment_names,
containing the solution summed over that vector compartment
}
\description{
sum time series of odin solution across compartments defined as vectors
}
