% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R, R/plotting.R
\name{thin_chain}
\alias{thin_chain}
\title{select evenly spaced iterations from MCMC chain}
\usage{
thin_chain(chain, n_samples)

thin_chain(chain, n_samples)
}
\arguments{
\item{chain}{data table containing MCMC iterations}

\item{n_samples}{numeric vector of length 1. number of samples to select}
}
\value{
data table of selected iterations

data table of selected iterations
}
\description{
select evenly spaced iterations from MCMC chain

select evenly spaced iterations from MCMC chain
}
