% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{thin_csv_or_rds}
\alias{thin_csv_or_rds}
\title{thins a csv or rds file, saves it with new name, optionally deletes old file}
\usage{
thin_csv_or_rds(filename, thin, save_as_rds, remove_after)
}
\arguments{
\item{filename}{character vector of length 1: with .csv or .rds extension}

\item{thin}{numeric vector of length 1 (must be in integer if not integer format):
thin every this many iterations}

\item{save_as_rds}{logical vector of length 1: if TRUE, save thinned chain as
rds regardless of extension of filename; if FALSE, save as original extension}

\item{remove_after}{logical vector of length 1: remove old file if TRUE}
}
\description{
thins a csv or rds file, saves it with new name, optionally deletes old file
}
