% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pars.R
\name{transform_pars_wrapper_two_strain}
\alias{transform_pars_wrapper_two_strain}
\title{create a closure for model 1 which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood}
\usage{
transform_pars_wrapper_two_strain(parTab, strains)
}
\arguments{
\item{parTab}{data frame containing priors}
}
\value{
a function which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
\description{
create a closure for model 1 which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
