% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{try_read}
\alias{try_read}
\title{reads rds or csv with tryCatch}
\usage{
try_read(filename, rds, ...)
}
\arguments{
\item{filename}{character vector of length 1: with or without extension}
}
\value{
the read data frame
}
\description{
reads rds or csv with tryCatch
}
\details{
rationale for existence: if it's not known beforehand whether we're reading
an rds or csv, may pass nonsensical arguments, so ignore if these are causing an error
}
