% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_script.R
\name{write_script}
\alias{write_script}
\title{write script to run job on cluster}
\usage{
write_script(R_filename)
}
\arguments{
\item{R_filename}{string ending in .R with R script to be run}
}
\value{
file with same name as R_filename but without extension
}
\description{
write script to run job on cluster
}
