% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_script.R
\name{write_script_wrapper}
\alias{write_script_wrapper}
\title{write scripts for omicron and delta on hNECs and Calu3}
\usage{
write_script_wrapper(base_name)
}
\arguments{
\item{base_name}{string ending in .R which forms basis of filename}
}
\value{
scripts
}
\description{
write scripts for omicron and delta on hNECs and Calu3
}
